+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - Quality metrics have added mask option
   - Bug fix for norm of CompositionOperator
   - Functions updated to use sapyb
+  - Fix KullbackLeibler numba gradient ignoring mask
   - show1D slice_list parameter can now be of type tuple 
 
 * 23.1.0